Created global confirm dialog

+++ lib/feature/inventory/presentation/widgets/category_form.dart
@@ -1,15 +1,14 @@
 // ignore_for_file: use_build_context_synchronously
-
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:tienda_pos/core/constant/app_colors.dart';
+import 'package:tienda_pos/core/constant/ui.dart';
 import 'package:tienda_pos/core/state/data_state.dart';
 import 'package:tienda_pos/core/styles/button_custom_styles.dart';
 import 'package:tienda_pos/core/styles/text_field_styles.dart';
-import 'package:tienda_pos/core/utils/ui/snackbar.dart';
+import 'package:tienda_pos/core/widgets/dialog.dart';
+import 'package:tienda_pos/core/widgets/snackbar.dart';
 import 'package:tienda_pos/feature/inventory/domain/entities/category/category_entity.dart';
 import 'package:tienda_pos/feature/inventory/presentation/view_models/category/category_notifier.dart';
 
@@ -54,7 +53,7 @@ class _CategoryFormState extends ConsumerState<CategoryForm> {
     final categoryState = ref.watch(categoryNortifierProvider);
 
     return Padding(
-      padding: const EdgeInsets.all(10),
+      padding: UI.dialog_padding,
       child: Form(
         key: _formKey,
         child: Column(
@@ -62,7 +61,7 @@ class _CategoryFormState extends ConsumerState<CategoryForm> {
           children: [
             const SizedBox(height: 20),
             const Text(
-              'Category Details',
+              'Category Name',
               style: TextStyle(fontWeight: FontWeight.w500, fontSize: 20),
             ),
             const SizedBox(height: 30),
@@ -72,29 +71,38 @@ class _CategoryFormState extends ConsumerState<CategoryForm> {
                 controller: nameController,
                 decoration: TextFieldStyles.decoration2(
                   InputDecoration(
-                    labelText: 'Category name',
                     suffixIcon: widget.categoryEntity?.id != null
                         ? IconButton(
                             icon: const Icon(Icons.delete),
                             color: AppColors.danger,
-                            onPressed: () async {
-                              final DataState result =
-                                  await categoryNotifier.delete();
-                              if (result.isSuccess) {
-                                if (widget.onDelete != null) {
-                                  widget.onDelete!();
-                                }
-                                Navigator.of(context).pop();
-                                showTopSnackbar(
-                                    context: context,
-                                    color: AppColors.success,
-                                    message: 'Category deleted successfully.');
-                              } else {
-                                showTopSnackbar(
-                                    context: context,
-                                    color: AppColors.danger,
-                                    message: result.error!);
-                              }
+                            onPressed: () {
+                              showConfirmDialog(
+                                context: context,
+                                confirmText: 'Delete',
+                                confirmColor: AppColors.danger,
+                                message:
+                                    'Confirm delete? This action cannot be undone.',
+                                onConfirm: () async {
+                                  final DataState result =
+                                      await categoryNotifier.delete();
+                                  if (result.isSuccess) {
+                                    if (widget.onDelete != null) {
+                                      widget.onDelete!();
+                                    }
+                                    Navigator.of(context).pop();
+                                    showTopSnackbar(
+                                        context: context,
+                                        color: AppColors.success,
+                                        message:
+                                            'Category deleted successfully.');
+                                  } else {
+                                    showTopSnackbar(
+                                        context: context,
+                                        color: AppColors.danger,
+                                        message: result.error!);
+                                  }
+                                },
+                              );
                             },
                           )
                         : null,
